+++ grafana-plugin/src/models/alertgroup/alertgroup.ts
@@ -7,10 +7,10 @@ import { makeRequest } from 'network';
 import { Mixpanel } from 'services/mixpanel';
 import { RootStore } from 'state';
 import { SelectOption } from 'state/types';
-import { showApiError, refreshPageError, openErrorNotification } from 'utils';
+import { openErrorNotification, refreshPageError, showApiError } from 'utils';
 import LocationHelper from 'utils/LocationHelper';
 
-import { Alert, AlertAction, IncidentStatus } from './alertgroup.types';
+import { Alert, AlertAction, IncidentStatus, ResponseIRMPlan } from './alertgroup.types';
 
 export class AlertGroupStore extends BaseStore {
   @observable.shallow
@@ -68,6 +68,9 @@ export class AlertGroupStore extends BaseStore {
 
   @observable
   liveUpdatesPaused = false;
+
+  @observable
+  irmPlan: ResponseIRMPlan = undefined;
 
   constructor(rootStore: RootStore) {
     super(rootStore);
@@ -202,6 +205,10 @@ export class AlertGroupStore extends BaseStore {
       method: 'POST',
       data: { template_name, template_body },
     });
+  }
+
+  async fetchIRMPlan() {
+    this.irmPlan = await makeRequest(`/usage-limits`, { method: 'GET' });
   }
 
   // methods were moved from rootBaseStore.